requirements.txt
pyserial==3.4
texttable==1.6.1
influxdb
gatt==0.2.7
pyedimax==0.1